Deleted extra activities due to bottom nav and used fragments instead
