+++ app/src/main/AndroidManifest.xml
@@ -11,6 +11,9 @@
         android:supportsRtl="true"
         android:theme="@style/SplashScreen">
         <activity
+            android:name=".view.MainActivity"
+            android:exported="false" />
+        <activity
             android:name=".view.download.DownloadActivity"
             android:exported="false" />
         <activity
