--- app/src/main/java/com/example/demoapp/view/download/DownloadActivity.kt
@@ -1,25 +1,0 @@
-package com.example.demoapp.view.download
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import com.example.demoapp.R
-import com.example.demoapp.view.homeandactionmenu.HomeScreenFragment
-
-class DownloadActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_download)
-//        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
-//            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-//            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-//            insets
-//        }
-        supportFragmentManager
-            .beginTransaction()
-            .replace(R.id.fragment_container,  DownloadFragment.newInstance("",""))
-            .commit();
-    }
-}
