+++ app/src/main/java/com/example/demoapp/view/homeandactionmenu/NewReleasesActivity.kt
@@ -2,8 +2,6 @@ package com.example.demoapp.view.homeandactionmenu
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import com.example.demoapp.R
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -12,10 +10,9 @@ class NewReleasesActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_new_releases)
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-            insets
-        }
+        supportFragmentManager
+            .beginTransaction()
+            .replace(R.id.fragment_container,  NewReleasesFragment())
+            .commit();
     }
 }
